feat: 퀀트 항로 상점, 소수점 수량 거래, UserInfo 프로필, 버그 수정
- 퀀트 항로 상점(QuantStore) 백엔드+프론트 전체 구현 (CRUD, 구매/취소, 자동 매수/매도)
- int→double 수량 마이그레이션: 암호화폐 소수점 거래 지원 (Order, Holding, TradeRecord 등)
- 항로 구매 시 타겟 자산에 균등 분배 매수, 취소 시 항로 수량만 매도
- 보유 종목 테이블에 항로 수량 breakdown 표시
- UserInfo 투자 프로필 (자기소개, 투자성향, 관심종목 등) 백엔드+프론트 구현
- 캔들스틱 차트, 지정가 주문 스케줄러, 빗썸 WebSocket 실시간 시세
- QuantStoreController NPE 방지, OrderRequest @DecimalMin 검증, division-by-zero 방어
- UI: 고래 테마 일관성, 한글명(코드) 표시, "항해 중인 항로" 용어 통일

+++ backend/src/main/java/com/project/whalearc/market/dto/MarketPriceResponse.java
@@ -15,8 +15,8 @@ public class MarketPriceResponse {
     private AssetType assetType; // STOCK or CRYPTO
     private String symbol;       // 종목 코드 or 코인 심볼 (005930, BTC 등)
     private String name;         // 종목/코인 이름
-    private long price;          // 현재가
-    private long change;         // 전일 대비 절대값
+    private double price;        // 현재가
+    private double change;       // 전일 대비 절대값
     private double changeRate;   // 전일 대비 %
     private long volume;         // 거래량 (24h 등)
     private String market;       // KRX, BITHUMB_KRW 등
